Fix: Budget Control Event
